Added a base class to provide common functionality for all math instructions.
