+++ CS6502.Core/Instructions/Math/AND.cs
@@ -2,7 +2,7 @@
 
 namespace CS6502.Core
 {
-    internal class AND : InstructionBase
+    internal class AND : MathInstructionBase
     {
         public static AND CreateAND(AddressingMode addressingMode)
         {
@@ -37,274 +37,6 @@ namespace CS6502.Core
             }
         }
 
-        public override CpuMicroCode Execute(
-            SignalEdge signalEdge,
-            int instructionCycle,
-            StatusRegister status,
-            bool wasPageBoundaryCrossed)
-        {
-            if (AddressingMode == AddressingMode.Immediate)
-            {
-                return Immediate(signalEdge, instructionCycle);
-            }
-            else if (AddressingMode == AddressingMode.ZeroPage ||
-                     AddressingMode == AddressingMode.ZeroPageX)
-            {
-                return ZeroPage(signalEdge, instructionCycle);
-            }
-            else if (AddressingMode == AddressingMode.Absolute ||
-                     AddressingMode == AddressingMode.AbsoluteX ||
-                     AddressingMode == AddressingMode.AbsoluteY)
-            {
-                return Absolute(signalEdge, instructionCycle, wasPageBoundaryCrossed);
-            }
-            else if (AddressingMode == AddressingMode.XIndirect ||
-                     AddressingMode == AddressingMode.IndirectY)
-            {
-                return Indirect(signalEdge, instructionCycle, wasPageBoundaryCrossed);
-            }
-            else
-            {
-                throw new ArgumentException($"AND does not support {AddressingMode.ToString()} addressing mode");
-            }
-        }
-
-        private CpuMicroCode Immediate(SignalEdge signalEdge, int instructionCycle)
-        {
-            if (signalEdge == SignalEdge.FallingEdge)
-            {
-                IsInstructionComplete = true;
-
-                if (instructionCycle == 2)
-                {
-                    return
-                        new CpuMicroCode(
-                            MicroCodeInstruction.AND,
-                            MicroCodeInstruction.IncrementPC,
-                            MicroCodeInstruction.TransferPCToPCS
-                        );
-                }
-            }
-
-            return new CpuMicroCode();
-        }
-
-        private CpuMicroCode ZeroPage(SignalEdge signalEdge, int instructionCycle)
-        {
-            int startingCycle = 2;
-            if (AddressingMode == AddressingMode.ZeroPageX)
-            {
-                startingCycle = 3;
-            }
-
-            if (signalEdge == SignalEdge.FallingEdge)
-            {
-                if (instructionCycle == startingCycle)
-                {
-                    if (AddressingMode == AddressingMode.ZeroPage)
-                    {
-                        return
-                            new CpuMicroCode(
-                               MicroCodeInstruction.TransferZPDataToAB,
-                               MicroCodeInstruction.LatchDataIntoDIL,
-                               MicroCodeInstruction.SetToRead,
-                               MicroCodeInstruction.IncrementPC
-                            );
-                    }
-                    else if (AddressingMode == AddressingMode.ZeroPageX)
-                    {
-                        return
-                            new CpuMicroCode(
-                               MicroCodeInstruction.LatchDataIntoDIL,
-                               MicroCodeInstruction.SetToRead,
-                               MicroCodeInstruction.IncrementABByX_NoCarry
-                            );
-                    }
-                }
-                if (instructionCycle == startingCycle + 1)
-                {
-                    IsInstructionComplete = true;
-
-                    return
-                        new CpuMicroCode(
-                            MicroCodeInstruction.AND,
-                            MicroCodeInstruction.TransferPCToPCS
-                        );
-                }
-            }
-            else
-            {
-                if (instructionCycle == startingCycle)
-                {
-                    return
-                        new CpuMicroCode(
-                            MicroCodeInstruction.TransferPCToPCS
-                        );
-                }
-            }
-
-            return new CpuMicroCode();
-        }
-
-        private CpuMicroCode Absolute(
-            SignalEdge signalEdge, 
-            int instructionCycle, 
-            bool wasPageBoundaryCrossed)
-        {
-            int startingCycle = 3;
-            if (AddressingMode == AddressingMode.AbsoluteX ||
-                AddressingMode == AddressingMode.AbsoluteY)
-            {
-                startingCycle = 4;
-            }
-
-            if (signalEdge == SignalEdge.FallingEdge)
-            {
-                if (instructionCycle == startingCycle)
-                {
-                    CpuMicroCode cpuMicroCode =
-                        new CpuMicroCode(
-                            MicroCodeInstruction.SetToRead
-                        );
-
-                    if (AddressingMode == AddressingMode.Absolute)
-                    {
-                        cpuMicroCode.Add(MicroCodeInstruction.TransferDILToPCHS);
-                        cpuMicroCode.Add(MicroCodeInstruction.TransferPCSToAddressBus);
-                        cpuMicroCode.Add(MicroCodeInstruction.IncrementPC);
-                    }
-
-                    if (AddressingMode == AddressingMode.AbsoluteX ||
-                        AddressingMode == AddressingMode.AbsoluteY)
-                    {
-                        if (wasPageBoundaryCrossed)
-                        {
-                            cpuMicroCode.Add(MicroCodeInstruction.IncrementABH);
-                            cpuMicroCode.Add(MicroCodeInstruction.ClearPageBoundaryCrossed);
-                        }
-                        else
-                        {
-                            IsInstructionComplete = true;
-                            cpuMicroCode.Add(MicroCodeInstruction.AND);
-                            cpuMicroCode.Add(MicroCodeInstruction.TransferPCToPCS);
-                        }
-                    }
-
-                    return cpuMicroCode;
-                }
-                else if (instructionCycle == startingCycle + 1)
-                {
-                    IsInstructionComplete = true;
-
-                    return
-                        new CpuMicroCode(
-                            MicroCodeInstruction.AND,
-                            MicroCodeInstruction.TransferPCToPCS
-                        );
-                }
-            }
-            else
-            {
-                if (instructionCycle == startingCycle)
-                {
-                    return
-                        new CpuMicroCode(
-                            MicroCodeInstruction.LatchDataIntoDIL
-                        );
-                }
-            }
-
-            return new CpuMicroCode();
-        }
-
-        private CpuMicroCode Indirect(
-            SignalEdge signalEdge, 
-            int instructionCycle,
-            bool wasPageBoundaryCrossed)
-        {
-            int startingCycle = 4;
-            if (AddressingMode == AddressingMode.XIndirect ||
-                AddressingMode == AddressingMode.IndirectY)
-            {
-                startingCycle = 5;
-            }
-
-            if (signalEdge == SignalEdge.FallingEdge)
-            {
-                if (AddressingMode == AddressingMode.XIndirect)
-                {
-                    if (instructionCycle == startingCycle)
-                    {
-                        if (AddressingMode == AddressingMode.XIndirect)
-                        {
-                            return
-                                new CpuMicroCode(
-                                    MicroCodeInstruction.SetToRead,
-                                    MicroCodeInstruction.TransferDILToPCHS,
-                                    MicroCodeInstruction.TransferPCSToAddressBus
-                                );
-                        }
-                    }
-                    else if (instructionCycle == startingCycle + 1)
-                    {
-                        IsInstructionComplete = true;
-
-                        return
-                            new CpuMicroCode(
-                                MicroCodeInstruction.AND,
-                                MicroCodeInstruction.TransferPCToPCS
-                            );
-                    }
-                }
-                else if (AddressingMode == AddressingMode.IndirectY)
-                {
-                    if (instructionCycle == startingCycle)
-                    {
-                        if (wasPageBoundaryCrossed)
-                        {
-                            return
-                                new CpuMicroCode(
-                                    MicroCodeInstruction.IncrementABH,
-                                    MicroCodeInstruction.ClearPageBoundaryCrossed
-                                );
-                        }
-                        else
-                        {
-                            IsInstructionComplete = true;
-
-                            return
-                                new CpuMicroCode(
-                                    MicroCodeInstruction.AND,
-                                    MicroCodeInstruction.TransferPCToPCS
-                                );
-                        }
-                    }
-                    else if (instructionCycle == startingCycle + 1)
-                    {
-                        IsInstructionComplete = true;
-
-                        return
-                            new CpuMicroCode(
-                                MicroCodeInstruction.AND,
-                                MicroCodeInstruction.TransferPCToPCS
-                            );
-                    }
-                }
-            }
-            else
-            {
-                if (instructionCycle == 3)
-                {
-                    return
-                        new CpuMicroCode(
-                            MicroCodeInstruction.LatchDataIntoDIL
-                        );
-                }
-            }
-
-            return new CpuMicroCode();
-        }
-
         private AND(
             byte opcode,
             AddressingMode addressingMode)
@@ -312,7 +44,8 @@ namespace CS6502.Core
             base(
                 "AND",
                 opcode,
-                addressingMode)
+                addressingMode,
+                MicroCodeInstruction.AND)
         {
         }
     }
